--- src/Exceptionless.Web/ClientApp/src/lib/components/data-table/data-table-checkbox.svelte
@@ -1,8 +1,0 @@
-<script lang="ts">
-	import { Checkbox } from '$comp/ui/checkbox';
-	import type { Writable } from 'svelte/store';
-
-	export let checked: Writable<boolean>;
-</script>
-
-<Checkbox bind:checked={$checked} {...$$restProps} />
