--- src/Exceptionless.Web/ClientApp/src/lib/components/data-table/data-table-row-actions.svelte
@@ -1,67 +1,0 @@
-<script lang="ts">
-	import { DotsHorizontal } from 'radix-icons-svelte';
-	import { Button } from '$comp/ui/button';
-	import * as DropdownMenu from '$comp/ui/dropdown-menu';
-
-	type Task = {
-		id: string;
-		title: string;
-		status: string;
-		label: string;
-		priority: string;
-	};
-
-	const labels = [
-		{
-			value: 'bug',
-			label: 'Bug'
-		},
-		{
-			value: 'feature',
-			label: 'Feature'
-		},
-		{
-			value: 'documentation',
-			label: 'Documentation'
-		}
-	];
-
-	export let row: Task;
-	const task = row; // taskSchema.parse(row);
-</script>
-
-<DropdownMenu.Root>
-	<DropdownMenu.Trigger asChild let:builder>
-		<Button
-			variant="ghost"
-			builders={[builder]}
-			class="flex h-8 w-8 p-0 data-[state=open]:bg-muted"
-		>
-			<DotsHorizontal class="w-4 h-4" />
-			<span class="sr-only">Open menu</span>
-		</Button>
-	</DropdownMenu.Trigger>
-	<DropdownMenu.Content class="w-[160px]" align="end">
-		<DropdownMenu.Item>Edit</DropdownMenu.Item>
-		<DropdownMenu.Item>Make a copy</DropdownMenu.Item>
-		<DropdownMenu.Item>Favorite</DropdownMenu.Item>
-		<DropdownMenu.Separator />
-		<DropdownMenu.Sub>
-			<DropdownMenu.SubTrigger>Labels</DropdownMenu.SubTrigger>
-			<DropdownMenu.SubContent>
-				<DropdownMenu.RadioGroup value={task.label}>
-					{#each labels as label}
-						<DropdownMenu.RadioItem value={label.value}>
-							{label.label}
-						</DropdownMenu.RadioItem>
-					{/each}
-				</DropdownMenu.RadioGroup>
-			</DropdownMenu.SubContent>
-		</DropdownMenu.Sub>
-		<DropdownMenu.Separator />
-		<DropdownMenu.Item>
-			Delete
-			<DropdownMenu.Shortcut>⌘⌫</DropdownMenu.Shortcut>
-		</DropdownMenu.Item>
-	</DropdownMenu.Content>
-</DropdownMenu.Root>
