+++ src/Exceptionless.Web/ClientApp/src/lib/components/data-table/index.ts
@@ -1,10 +1,29 @@
-export { default as DataTableCheckbox } from './data-table-checkbox.svelte';
-export { default as DataTableTitleCell } from './data-table-title-cell.svelte';
-export { default as DataTableStatusCell } from './data-table-status-cell.svelte';
-export { default as DataTableRowActions } from './data-table-row-actions.svelte';
-export { default as DataTablePriorityCell } from './data-table-priority-cell.svelte';
-export { default as DataTableColumnHeader } from './data-table-column-header.svelte';
-export { default as DataTableToolbar } from './data-table-toolbar.svelte';
-export { default as DataTablePagination } from './data-table-pagination.svelte';
-export { default as DataTableViewOptions } from './data-table-view-options.svelte';
-export { default as DataTableFacetedFilter } from './data-table-faceted-filter.svelte';
+import Root from './data-table.svelte';
+import Body from './data-table-body.svelte';
+import Toolbar from './data-table-toolbar.svelte';
+import FacetedFilter from './data-table-faceted-filter.svelte';
+import Pagination from './data-table-pagination.svelte';
+
+import PriorityCell from './data-table-priority-cell.svelte';
+import StatusCell from './data-table-status-cell.svelte';
+import TitleCell from './data-table-title-cell.svelte';
+
+export {
+	Root,
+	Body,
+	Toolbar,
+	FacetedFilter,
+	Pagination,
+	PriorityCell,
+	StatusCell,
+	TitleCell,
+	//
+	Root as DataTable,
+	Body as DataTableBody,
+	Toolbar as DataTableToolbar,
+	FacetedFilter as DataTableFacetedFilter,
+	Pagination as DataTablePagination,
+	PriorityCell as DataTablePriorityCell,
+	StatusCell as DataTableStatusCell,
+	TitleCell as DataTableTitleCell
+};
